+++ fiddles/jquery/fiddle-0035-JqueryCSV/app.js
@@ -3,15 +3,16 @@
 
     $(document).ready(function () {
 
-        $(document).load( "dates.csv" );
+        $(document).load( "data.csv" );
 
         $(document).ajaxComplete(function (event, xhr, settings) {
-            if (settings.url === "dates.csv") {
+            if (settings.url === "data.csv") {
                 var hook = window.document.getElementById('fiddleHook'),
                     container = window.document.createElement('pre'),
                     data = $.csv.toObjects(xhr.responseText),
                     raw = JSON.stringify(data);
                 container.innerText = raw;
+                console.log(JSON.stringify(data));
                 hook.appendChild(container);
             }
         })