+++ CHANGELOG.md
@@ -2130,4 +2130,11 @@ Alternately, the _Daily Diary of my Programming Dreams_.
   
   1.  Added [fiddles/jquery/fiddle-0074-AgGridEnterprise](fiddles/jquery/fiddle-0074-AgGridEnterprise)
 
-* Added [fiddles/angular2-cli/fiddle-0019-EnterpriseAgGrid](fiddles/angular2-cli/fiddle-0019-EnterpriseAgGrid)
+
+### 297_angularCli_20 ~ [Issue 297](https://github.com/bradyhouse/house/issues/297)
+
+* **angular2-cli**
+
+  1.  Added [fiddles/angular2-cli/fiddle-0019-EnterpriseAgGrid](fiddles/angular2-cli/fiddle-0019-EnterpriseAgGrid)
+  2.  Added [fiddles/angular2-cli/fiddle-0020-EntAgGridPivot](fiddles/angular2-cli/fiddle-0020-EntAgGridPivot)
+